Cleaned up Sim Control Panel test suite
